move [NonError]QueryResponse to response module
