+++ scylla/src/response/mod.rs
@@ -1,1 +1,5 @@
+mod request_response;
 
+pub(crate) use request_response::{
+    NonErrorAuthResponse, NonErrorQueryResponse, NonErrorStartupResponse, QueryResponse,
+};
